replace carousel animations properties for an object
