+++ mlbusinesscomponents/src/main/java/com/mercadolibre/android/mlbusinesscomponents/components/touchpoint/domain/model/cover_carousel/response/CarouselAnimationInterface.java
@@ -1,0 +1,10 @@
+package com.mercadolibre.android.mlbusinesscomponents.components.touchpoint.domain.model.cover_carousel.response;
+
+public interface CarouselAnimationInterface {
+
+    boolean getAlphaAnimation();
+
+    boolean getScaleAnimation();
+
+    boolean getPressAnimation();
+}
