+++ mlbusinesscomponents/src/main/java/com/mercadolibre/android/mlbusinesscomponents/components/touchpoint/domain/model/cover_carousel/response/CoverCarouselInterface.java
@@ -8,11 +8,7 @@ public interface CoverCarouselInterface {
 
     HeaderInterface getHeader();
 
-    boolean getAlphaAnimation();
-
-    boolean getScaleAnimation();
-
-    boolean getPressAnimation();
+    CarouselAnimationInterface getCarouselAnimation();
 
     List<CoverCardInterface> getItems();
 }
